refactor(#438): split admin/travel.js into focused sub-modules
513-line monolith extracted into travel/ sub-module directory:
- messages.js  — shared setMessage via createMessenger
- media.js     — pendingFiles/existingMedia state + renderMediaList
- geocode.js   — Leaflet map, Nominatim forward/reverse geocode, initGeocode
- exif.js      — EXIF GPS/date autofill using exifr
- list.js      — loadAll, loadForEdit, togglePublish, deleteMemory
- form.js      — clearForm, initForm (file-change, submit, cancel, clear)

admin/travel.js is now an 8-line orchestrator that imports and calls
initTravel() → loadAll() + initForm() + initGeocode(). No behaviour changes.
