+++ resources/js/admin/travel/messages.js
@@ -1,0 +1,3 @@
+import { createMessenger } from '../../utils/messenger.js';
+
+export const setMessage = createMessenger('travel-message');
